AlternativesCheck: Fix .conf files regex
The previous regex matches any path in /usr/share/libalternatives/ that ends
with "conf" so it was crashing when trying to parse an rpm file that contains a
binary "ldaptor-ldap2dhcpconf" with a libalternatives configuration.

This patch fixes this issue making the regex more specific so it doesn't catch
any folder, just ".conf" files.